+++ common/core/src/test/java/is/codion/common/value/ValueTest.java
@@ -88,7 +88,6 @@ public class ValueTest {
 				assertNotNull(data);
 			}
 		});
-		assertFalse(intValue.set(42));
 		assertEquals(0, eventCounter.get());
 		intValue.set(20);
 		assertEquals(1, eventCounter.get());